pyproject.toml
[project]
name = "autonomous-qa-agents"
version = "0.1.0"
description = "Agentic QA automation system with LLMs, CrewAI, and Confluence integration"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "crewai",
    "python-dotenv",
    "requests",
    "playwright",
    "openai",
    "langchain",
    "streamlit",
    "groq"
]
